+++ src/app/homeFavorite/home-favorite.resolvers.ts
@@ -1,13 +1,17 @@
-import {ParseIntPipe, UseGuards} from '@nestjs/common';
+import {UnauthorizedException, UseGuards} from '@nestjs/common';
 import {Args, Mutation, Query, Resolver, Subscription} from '@nestjs/graphql';
 import {PubSub} from 'graphql-subscriptions';
 import {HomeFavoriteService} from './home-favorite.service';
-import {HomeFavorite} from '../graphql.schema';
+import {Home, HomeFavorite} from '../graphql.schema';
 import {CreateHomeFavoriteDto, DeleteHomeFavoriteDto} from './dto';
+import {GraphqlGuard} from '../_helpers';
+import {User as CurrentUser} from '../_helpers/graphql/user.decorator';
+import {UserEntity as User} from '../user/entity';
 
 const pubSub = new PubSub();
 
 @Resolver('HomeFavorite')
+@UseGuards(GraphqlGuard)
 export class HomeFavoriteResolvers {
 	constructor(private readonly homeFavoriteService: HomeFavoriteService) {
 	}
@@ -18,17 +22,23 @@ export class HomeFavoriteResolvers {
 	}
 
 	@Mutation('createHomeFavorite')
-	async create(@Args('createHomeFavoriteInput') args: CreateHomeFavoriteDto): Promise<HomeFavorite> {
+	async create(@CurrentUser() user: User, @Args('createHomeFavoriteInput') args: CreateHomeFavoriteDto): Promise<HomeFavorite> {
 		const createdHomeFavorite = await this.homeFavoriteService.create(args);
+		args.homeFavoriteUserId = user.id;
 		pubSub.publish('homeFavoriteCreated', {homeCreatedFavorite: createdHomeFavorite});
 		return createdHomeFavorite;
 	}
 
 	@Mutation('deleteHomeFavorite')
-	async delete(@Args('deleteHomeFavoriteInput') args: DeleteHomeFavoriteDto): Promise<HomeFavorite> {
-		const deletedHomeFavorite = await this.homeFavoriteService.delete(args.id);
-		pubSub.publish('homeFavoriteDeleted', {homeFavoriteDeleted: deletedHomeFavorite});
-		return deletedHomeFavorite;
+	async delete(@CurrentUser() user: User, @Args('deleteHomeFavoriteInput') args: DeleteHomeFavoriteDto): Promise<HomeFavorite> {
+		const homeFavoritesToDelete: Home = await this.homeFavoriteService.findOneById(args.id);
+		if (homeFavoritesToDelete.owner === user.id) {
+			const deletedHomeFavorite = await this.homeFavoriteService.delete(args.id);
+			pubSub.publish('homeFavoriteDeleted', {homeFavoriteDeleted: deletedHomeFavorite});
+			return deletedHomeFavorite;
+		} else {
+			throw new UnauthorizedException();
+		}
 	}
 
 	@Subscription('homeFavoriteCreated')
